dataflow/requirements.txt
google-cloud-language
apache-beam[gcp]
tinybird-beam
